Separation of capabilities for initContainers
This allows a separate set of capabilities to be defined for initContainers which may be more or less restrictive than the capability set for containers that are not initContainers. This facilitates situations where you need higher privileges in an initContainer for the express purpose of doing pre-configuration in order to run a container for the main service with lesser capabilities.

Signed-off-by: Thomas Spear <tspear@conquestcyber.com>