fix(k8spsphostnetworkingports): null check container.ports
Previously, the `badContainers` section of the K8sNative validation
logic did not verify that the `ports` variable was set on a container
spec in a pod.  As this template's purpose is to verify that the
specified ports fall within a given range, we would expect that _no
ports_ is not a violation.

Signed-off-by: juliankatz <juliankatz@google.com>