fix(k8srequiredlabels): CEL broke when allowedRegex was empty
The CEL implementation of this policy always expected the `allowedRegex`
value to be set for a given `key` value in the contraint parameters.

Existing users of the template could (via rego) require only that a
given label key exist, but assert nothing about the value of that label.

This change makes the CEL implementation uphold that same contract.

Signed-off-by: juliankatz <juliankatz@google.com>