fix(k8spspreadonlyrootfilesystem): CEL support wildcard in exemptImages
Despite it being non-sensical to put a `*` in exemptImages (functionally
disabling the policy), this is supported in the existing rego
implementation of the template.

Thus, not doing it in the CEL implementation is an inconsistency and a
breaking change.

This PR upholds the contract by adding support for `*` as an
exemptImage.

Signed-off-by: juliankatz <juliankatz@google.com>