chore(k8spspallowprivilegeescalationcontainer): suite test confirming exemptImages
I recently found (#584) that some K8sNativeValidation implementations of
certain templates that iterate over and exempt containers by image had
a bug preventing the exemption logic from working.

The k8spspallowprivilegeescalationcontainer turns out not to have this
problem, as proved by the passing tests with the addition of an image
exemption suite test.

Signed-off-by: juliankatz <juliankatz@google.com>