Rego v1 and Regal compliance
This is rewrite of the entire Rego library for Rego v1 compliance, and
the introduction of Regal to identify issues and enforce good coding standards.

To not be too disruptive, all policies were first copied to a `v1` directory under
`src`, and no modifications to existing policies have been made. The Regal configuration
is also set to ignore anything but the new v1 directory.

The result is about 1200 lines less Rego in the v1 directory compared to the previous
versions.

There's still a bunch of practicalities to deal with here, I'm sure. Like how to integrate
this in the project CI, publishing, and so on. The templates are currently just copied from
the "old" directories, but should probably not follow the same versioning, and might need
something to identify them as Rego v1 policies. And so on, and so forth :)

But I figured a PR would be the best way to start that discussion, and for me to get some
directions for what to work on now.

The test.sh file have been updated to run the unit tests against the v1 policies too, and
they're all passing. I have not yet played with the Gator CLI tests though, so that might
be a good thing to look into as well.

Signed-off-by: Anders Eknert <anders@styra.com>